move Scala 2 specific pdt-intersection-summon progression test to shared tests since it compiles on Scala 3
