+++ izumi-reflect/izumi-reflect/src/test/scala-2/izumi/reflect/test/TagProgressionTest.scala
@@ -7,38 +7,7 @@ class TagProgressionTest extends SharedTagProgressionTest {
 
   "[progression] Tag (Scala 2)" should {
 
-    "progression test: combine intersection path-dependent intersection types with inner tags doesn't work yet (Scala 2 specific, generic type projection)" in {
-      trait PDT {
-        type T
-        implicit def tag: Tag[T]
-
-        def badCombine(that: PDT): Tag[T with that.T] = {
-          Tag[T with that.T]
-        }
-      }
-      broken {
-        assertCompiles(
-          """
-          trait PDT0 {
-            type T
-            implicit def tag: Tag[T]
-
-            def goodCombine(that: PDT): Tag[T with that.T] = {
-              import that.tag
-              Tag[T with that.T]
-            }
-          }"""
-        )
-      }
-      def PDT[U: Tag]: PDT = new PDT { type T = U; override val tag: Tag[U] = Tag[U] }
-
-      val badCombine = PDT[Int].badCombine(PDT[Unit])
-      broken {
-        assertSameStrict(badCombine.tag, Tag[Int with Unit].tag)
-      }
-    }
-
-    "progression test: type tags with bounds are not currently requested by the macro on Scala 2" in {
+    "progression test: type tags with bounds are not currently requested by the macro on Scala 2 (using Scala 2 specific HKTag)" in {
       val t = intercept[TestFailedException] {
         assertCompiles("""
         type `TagK<:Dep`[K[_ <: Dep]] = izumi.reflect.HKTag[ { type Arg[A <: Dep] = K[A] } ]
